move getNewVolume to item util and use everywhere
