+++ client/packages/inventory/src/Stocktake/DetailView/modal/StocktakeLineEdit/StocktakeLineEditTables.tsx
@@ -30,6 +30,7 @@ import {
   getCampaignOrProgramColumn,
   getDonorColumn,
   getLocationInputColumn,
+  getVolumePerPackFromVariant,
   ItemVariantInputCell,
   PackSizeEntryCell,
   ReasonOptionRowFragment,
@@ -212,21 +213,11 @@ export const BatchTable = ({
           <ItemVariantInputCell {...props} itemId={props.rowData.item.id} />
         ),
         setter: patch => {
-          const { packSize, itemVariant } = patch;
-
-          if (itemVariant) {
-            const packaging = itemVariant.packagingVariants.find(
-              p => p.packSize === packSize
-            );
-            // Item variants save volume in L, but it is saved in m3 everywhere else
-            update({
-              ...patch,
-              volumePerPack:
-                ((packaging?.volumePerUnit ?? 0) / 1000) * (packSize ?? 1),
-            });
-          } else {
-            update(patch);
-          }
+          update({
+            ...patch,
+            volumePerPack:
+              getVolumePerPackFromVariant(patch) ?? patch.volumePerPack,
+          });
         },
       });
     }
